Update ftp_installation.md
Is this correct? Also will they ever see it with an "s"?

 /www/httpdocs
/www/httpsdocs

All about the note:

I think the note on line 19 needs to be clearer because I don't quite understand it. 

We don't want people to get frustrated, so this note needs to be clearly defined and put well before the instructions to move all the folders and files. Comprende? :)

Also, the example should be consistent, so if the example is going to be myshop, they should be told to create a folder named "myshop".

I may want to look this over again at some point.