Update price-rules.md
since this is everywhere, I'm leaving it alone from now on, if you have a way to globally replace it:

If you would like to download this help file as a PDF, you can do so here. The PDF is comprehensive of all the content in the online help. 

Proposed replacement:

If you would like to download this help file as a PDF,  you can do so here. The PDF is a comprehensive guide that includes all the content in the online help.